self-org/requirements.txt
tensorflow 
Pillow 
matplotlib 
tqdm 
moviepy 
ffmpeg